+++ src/core/selection/style/api/index.ts
@@ -4,7 +4,7 @@
  * Copyright (c) 2013-2022 Valeriy Chupurnov. All rights reserved. https://xdsoft.net
  */
 
-export * from './toggle/toggle-css';
+export * from './toggle/toggle-attributes';
 export * from './toggle/toggle-ordered-list';
 export * from './element-has-same-style';
 export * from './extract';
@@ -12,10 +12,11 @@ export * from './finite-state-machine';
 export * from './get-suit-child';
 export * from './get-suit-parent';
 export * from './is-inside-invisible-element';
+export * from './is-same-attributes';
 export * from './is-normal-node';
 export * from './is-suit-element';
 export * from './toggle-commit-styles';
 export * from './unwrap-children';
 export * from './wrap-and-commit-style';
-export * from './wrap-ordered-list';
+export * from './wrap-list';
 export * from './wrap-unwrapped-text';