+++ CHANGELOG.md
@@ -9,6 +9,50 @@
 > -   :house: [Internal]
 > -   :nail_care: [Polish]
 
+## 3.24.1
+
+#### :rocket: New Feature
+
+-   Method `Select.applyStyle` marked as deprecated. Use `Select.commitStyle` instead.
+
+Before:
+
+```js
+jodit.select.applyStyle(
+	{ color: red },
+	{
+		element: 'strong'
+	}
+);
+```
+
+Now:
+
+```js
+jodit.s.commitStyle({
+	element: 'strong',
+	style: {
+		color: 'red'
+	}
+});
+```
+
+-   In the options of the `Select`.`commitStyle` method, the `attributes` property has been added, which allows you to
+    also set attributes when applying a style.
+
+```js
+jodit.s.commitStyle({
+	element: 'a',
+	attributes: {
+		href: 'https://stename.ru'
+	}
+});
+```
+
+Wraps the selected text into a link with the specified address.
+
+-   When inserting a url, if the text is selected, it will automatically be replaced with a link
+
 ## 3.23.3
 
 #### :rocket: New Feature
@@ -35,7 +79,7 @@ Jodit.make('.editor', {
 		}
 	]
 });
-```
+````
 
 ## 3.23.2
 
@@ -1824,11 +1868,11 @@ Related with https://github.com/xdan/jodit/issues/574. In some cases need to lim
 -   @property {IUIOption[]} link.selectOptionsClassName=[] The list of the option for the select (to use with
     modeClassName="select")
 -   ex: [
--                                                                                                                                                                                                                                                                                                                                                                                                                                           	{ value: "", text: "" },
--                                                                                                                                                                                                                                                                                                                                                                                                                                           	{ value: "val1", text: "text1" },
--                                                                                                                                                                                                                                                                                                                                                                                                                                           	{ value: "val2", text: "text2" },
--                                                                                                                                                                                                                                                                                                                                                                                                                                           	{ value: "val3", text: "text3" }
--                                                                                                                                                                                                                                                                                                                                                                                                                                           ]
+-                                                                                                                                                                                                                                                                                                                                                                                                                                                         	{ value: "", text: "" },
+-                                                                                                                                                                                                                                                                                                                                                                                                                                                         	{ value: "val1", text: "text1" },
+-                                                                                                                                                                                                                                                                                                                                                                                                                                                         	{ value: "val2", text: "text2" },
+-                                                                                                                                                                                                                                                                                                                                                                                                                                                         	{ value: "val3", text: "text3" }
+-                                                                                                                                                                                                                                                                                                                                                                                                                                                         ]
     PR: https://github.com/xdan/jodit/pull/577 Thanks @s-renier-taonix-fr
 
 ##### New option `statusbar: boolean = true`